Function pallet_nfts::pallet::dispatchables::approve_transfer

pub fn approve_transfer<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    item: T::ItemId,
    delegate: <<T as SystemConfig>::Lookup as StaticLookup>::Source,
    maybe_deadline: Option<BlockNumberFor<T>>,
)

Approve an item to be transferred by a delegated third-party account.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the item.

collection: The collection of the item to be approved for delegated transfer.

item: The item to be approved for delegated transfer.

delegate: The account to delegate permission to transfer the item.

maybe_deadline: Optional deadline for the approval. Specified by providing the number of blocks after which the approval will expire

Emits TransferApproved on success.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::approve_transfer.