pallet_nfts::pallet::dispatchables

Function lock_collection

pub fn lock_collection<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    lock_settings: CollectionSettings,
)

Disallows specified settings for the whole collection.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection to be locked.

lock_settings: The settings to be locked.

Note: it’s possible to only lock(set) the setting, but not to unset it.

Emits CollectionLocked.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::lock_collection.