Function pallet_nfts::pallet::dispatchables::mint_pre_signed

pub fn mint_pre_signed<T: Config<I>, I: 'static>(
    mint_data: Box<PreSignedMint<<T as Config<I>>::CollectionId, <T as Config<I>>::ItemId, <T as SystemConfig>::AccountId, BlockNumberFor<T>, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>>,
    signature: T::OffchainSignature,
    signer: T::AccountId,
)

Mint an item by providing the pre-signed approval.

Origin must be Signed.

mint_data: The pre-signed approval that consists of the information about the item, its metadata, attributes, who can mint it (None for anyone) and until what block number.

signature: The signature of the data object.

signer: The data object’s signer. Should be an Issuer of the collection.

Emits Issued on success. Emits AttributeSet if the attributes were provided. Emits ItemMetadataSet if the metadata was not empty.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::mint_pre_signed.