pallet_nfts::pallet::dispatchables

Function redeposit

pub fn redeposit<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    items: Vec<T::ItemId>,
)

Re-evaluate the deposits on some items.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection of the items to be reevaluated.

items: The items of the collection whose deposits will be reevaluated.

NOTE: This exists as a best-effort function. Any items which are unknown or in the case that the owner account does not have reservable funds to pay for a deposit increase are ignored. Generally the owner isn’t going to call this on items whose existing deposit is less than the refreshed deposit as it would only cost them, so it’s of little consequence.

It will still return an error in the case that the collection is unknown or the signer is not permitted to call it.

Weight: O(items.len())

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::redeposit.