pallet_nfts::pallet::dispatchables

Function set_accept_ownership

pub fn set_accept_ownership<T: Config<I>, I: 'static>(
    maybe_collection: Option<T::CollectionId>,
)

Set (or reset) the acceptance of ownership for a particular account.

Origin must be Signed and if maybe_collection is Some, then the signer must have a provider reference.

maybe_collection: The identifier of the collection whose ownership the signer is willing to accept, or if None, an indication that the signer is willing to accept no ownership transferal.

Emits OwnershipAcceptanceChanged.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_accept_ownership.