Function pallet_nfts::pallet::dispatchables::set_attributes_pre_signed

pub fn set_attributes_pre_signed<T: Config<I>, I: 'static>(
    data: PreSignedAttributes<<T as Config<I>>::CollectionId, <T as Config<I>>::ItemId, <T as SystemConfig>::AccountId, BlockNumberFor<T>>,
    signature: T::OffchainSignature,
    signer: T::AccountId,
)

Set attributes for an item by providing the pre-signed approval.

Origin must be Signed and must be an owner of the data.item.

data: The pre-signed approval that consists of the information about the item, attributes to update and until what block number.

signature: The signature of the data object.

signer: The data object’s signer. Should be an Admin of the collection for the CollectionOwner namespace.

Emits AttributeSet for each provided attribute. Emits ItemAttributesApprovalAdded if the approval wasn’t set before. Emits PreSignedAttributesSet on success.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_attributes_pre_signed.