Function pallet_nfts::pallet::dispatchables::set_collection_metadata

pub fn set_collection_metadata<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    data: BoundedVec<u8, T::StringLimit>
)

Set the metadata for a collection.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

If the origin is Signed, then funds of signer are reserved according to the formula: MetadataDepositBase + DepositPerByte * data.len taking into account any already reserved funds.

collection: The identifier of the item whose metadata to update.

data: The general information of this item. Limited in length by StringLimit.

Emits CollectionMetadataSet.

Weight: O(1)

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the set_collection_metadata variant of the pallet Call enum. You should not attempt to call this function directly.