Struct pallet_nfts::pallet::Pallet

pub struct Pallet<T, I = ()>(_);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn owner( collection: T::CollectionId, item: T::ItemId ) -> Option<T::AccountId>

Get the owner of the item, if the item exists.

pub fn collection_owner(collection: T::CollectionId) -> Option<T::AccountId>

Get the owner of the collection, if the collection exists.

pub fn validate_signature( data: &Vec<u8>, signature: &T::OffchainSignature, signer: &T::AccountId ) -> DispatchResult

Validates the signature of the given data with the provided signer’s account ID.

Errors

This function returns a WrongSignature error if the signature is invalid or the verification process fails.

pub fn set_next_id(id: T::CollectionId)

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn construct_attribute_key( key: Vec<u8> ) -> Result<BoundedVec<u8, T::KeyLimit>, DispatchError>

A helper method to construct an attribute’s key.

Errors

This function returns an IncorrectData error if the provided attribute key is too long.

pub fn construct_attribute_value( value: Vec<u8> ) -> Result<BoundedVec<u8, T::ValueLimit>, DispatchError>

A helper method to construct an attribute’s value.

Errors

This function returns an IncorrectData error if the provided value is too long.

pub fn has_system_attribute( collection: &T::CollectionId, item: &T::ItemId, attribute_key: PalletAttributes<T::CollectionId> ) -> Result<bool, DispatchError>

A helper method to check whether a system attribute is set for a given item.

Errors

This function returns an IncorrectData error if the provided pallet attribute is too long.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn do_create_collection( collection: T::CollectionId, owner: T::AccountId, admin: T::AccountId, config: CollectionConfig<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, <T as Config<I>>::CollectionId>, deposit: <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, event: Event<T, I> ) -> DispatchResult

Create a new collection with the given collection, owner, admin, config, deposit, and event.

This function creates a new collection with the provided parameters. It reserves the required deposit from the owner’s account, sets the collection details, assigns admin roles, and inserts the provided configuration. Finally, it emits the specified event upon success.

Errors

This function returns a CollectionIdInUse error if the collection ID is already in use.

pub fn do_destroy_collection( collection: T::CollectionId, witness: DestroyWitness, maybe_check_owner: Option<T::AccountId> ) -> Result<DestroyWitness, DispatchError>

Destroy the specified collection with the given collection, witness, and maybe_check_owner.

This function destroys the specified collection if it exists and meets the necessary conditions. It checks the provided witness against the actual collection details and removes the collection along with its associated metadata, attributes, and configurations. The necessary deposits are returned to the corresponding accounts, and the roles and configurations for the collection are cleared. Finally, it emits the Destroyed event upon successful destruction.

Errors

This function returns a dispatch error in the following cases:

If the collection ID is not found (UnknownCollection).

If the provided maybe_check_owner does not match the actual owner (NoPermission).

If the collection is not empty (contains items) (CollectionNotEmpty).

If the witness does not match the actual collection details (BadWitness).

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn do_mint( collection: T::CollectionId, item: T::ItemId, maybe_depositor: Option<T::AccountId>, mint_to: T::AccountId, item_config: ItemConfig, with_details_and_config: impl FnOnce(&CollectionDetails<<T as SystemConfig>::AccountId, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>, &CollectionConfig<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, <T as Config<I>>::CollectionId>) -> DispatchResult ) -> DispatchResult

Mint a new unique item with the given collection, item, and other minting configuration details.

This function performs the minting of a new unique item. It checks if the item does not already exist in the given collection, and if the max supply limit (if configured) is not reached. It also reserves the required deposit for the item and sets the item details accordingly.

Errors

This function returns a dispatch error in the following cases:

If the collection ID is invalid (UnknownCollection).

If the item already exists in the collection (AlreadyExists).

If the item configuration already exists (InconsistentItemConfig).

If the max supply limit (if configured) for the collection is reached (MaxSupplyReached).

If any error occurs in the with_details_and_config closure.

pub fn do_burn( collection: T::CollectionId, item: T::ItemId, with_details: impl FnOnce(&ItemDetails<<T as SystemConfig>::AccountId, ItemDeposit<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, <T as SystemConfig>::AccountId>, BoundedBTreeMap<<T as SystemConfig>::AccountId, Option<BlockNumberFor<T>>, <T as Config<I>>::ApprovalsLimit>>) -> DispatchResult ) -> DispatchResult

Burns the specified item with the given collection, item, and with_details.

Errors

This function returns a dispatch error in the following cases:

If the collection ID is invalid (UnknownCollection).

If the item is locked (ItemLocked).

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn construct_metadata( metadata: Vec<u8> ) -> Result<BoundedVec<u8, T::StringLimit>, DispatchError>

A helper method to construct metadata.

Errors

This function returns an IncorrectMetadata dispatch error if the provided metadata is too long.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn group_roles_by_account( input: Vec<(T::AccountId, CollectionRole)> ) -> Vec<(T::AccountId, CollectionRoles)>

Groups provided roles by account, given one account could have multiple roles.

input: A vector of (Account, Role) tuples.

Returns a grouped vector of (Account, Roles) tuples.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn do_transfer( collection: T::CollectionId, item: T::ItemId, dest: T::AccountId, with_details: impl FnOnce(&CollectionDetails<<T as SystemConfig>::AccountId, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>, &mut ItemDetails<<T as SystemConfig>::AccountId, ItemDeposit<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, <T as SystemConfig>::AccountId>, BoundedBTreeMap<<T as SystemConfig>::AccountId, Option<BlockNumberFor<T>>, <T as Config<I>>::ApprovalsLimit>>) -> DispatchResult ) -> DispatchResult

Transfer an NFT to the specified destination account.

collection: The ID of the collection to which the NFT belongs.

item: The ID of the NFT to transfer.

dest: The destination account to which the NFT will be transferred.

with_details: A closure that provides access to the collection and item details, allowing customization of the transfer process.

This function performs the actual transfer of an NFT to the destination account. It checks various conditions like item lock status and transferability settings for the collection and item before transferring the NFT.

Errors

This function returns a dispatch error in the following cases:

If the collection ID is invalid (UnknownCollection).

If the item ID is invalid (UnknownItem).

If the item is locked or transferring it is disabled (ItemLocked).

If the collection or item is non-transferable (ItemsNonTransferable).

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn create( origin: OriginFor<T>, admin: <<T as SystemConfig>::Lookup as StaticLookup>::Source, config: CollectionConfig<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, <T as Config<I>>::CollectionId> ) -> DispatchResult

Issue a new collection of non-fungible items from a public origin.

This new collection has no items initially and its owner is the origin.

The origin must be Signed and the sender must have sufficient funds free.

CollectionDeposit funds of sender are reserved.

Parameters:

admin: The admin of this collection. The admin is the initial address of each member of the collection’s admin team.

Emits Created event when successful.

Weight: O(1)

pub fn force_create( origin: OriginFor<T>, owner: <<T as SystemConfig>::Lookup as StaticLookup>::Source, config: CollectionConfig<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, <T as Config<I>>::CollectionId> ) -> DispatchResult

Issue a new collection of non-fungible items from a privileged origin.

This new collection has no items initially.

The origin must conform to ForceOrigin.

Unlike create, no funds are reserved.

owner: The owner of this collection of items. The owner has full superuser permissions over this item, but may later change and configure the permissions using transfer_ownership and set_team.

Emits ForceCreated event when successful.

Weight: O(1)

pub fn destroy( origin: OriginFor<T>, collection: T::CollectionId, witness: DestroyWitness ) -> DispatchResultWithPostInfo

Destroy a collection of fungible items.

The origin must conform to ForceOrigin or must be Signed and the sender must be the owner of the collection.

NOTE: The collection must have 0 items to be destroyed.

collection: The identifier of the collection to be destroyed.

witness: Information on the items minted in the collection. This must be correct.

Emits Destroyed event when successful.

Weight: O(m + c + a) where:

m = witness.item_metadatas

c = witness.item_configs

a = witness.attributes

pub fn mint( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, mint_to: <<T as SystemConfig>::Lookup as StaticLookup>::Source, witness_data: Option<MintWitness<T::ItemId, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>> ) -> DispatchResult

Mint an item of a particular collection.

The origin must be Signed and the sender must comply with the mint_settings rules.

collection: The collection of the item to be minted.

item: An identifier of the new item.

mint_to: Account into which the item will be minted.

witness_data: When the mint type is HolderOf(collection_id), then the owned item_id from that collection needs to be provided within the witness data object. If the mint price is set, then it should be additionally confirmed in the witness_data.

Note: the deposit will be taken from the origin and not the owner of the item.

Emits Issued event when successful.

Weight: O(1)

pub fn force_mint( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, mint_to: <<T as SystemConfig>::Lookup as StaticLookup>::Source, item_config: ItemConfig ) -> DispatchResult

Mint an item of a particular collection from a privileged origin.

The origin must conform to ForceOrigin or must be Signed and the sender must be the Issuer of the collection.

collection: The collection of the item to be minted.

item: An identifier of the new item.

mint_to: Account into which the item will be minted.

item_config: A config of the new item.

Emits Issued event when successful.

Weight: O(1)

pub fn burn( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId ) -> DispatchResult

Destroy a single item.

The origin must conform to ForceOrigin or must be Signed and the signing account must be the owner of the item.

collection: The collection of the item to be burned.

item: The item to be burned.

Emits Burned.

Weight: O(1)

pub fn transfer( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, dest: <<T as SystemConfig>::Lookup as StaticLookup>::Source ) -> DispatchResult

Move an item from the sender account to another.

Origin must be Signed and the signing account must be either:

the Owner of the item;

the approved delegate for the item (in this case, the approval is reset).

Arguments:

collection: The collection of the item to be transferred.

item: The item to be transferred.

dest: The account to receive ownership of the item.

Emits Transferred.

Weight: O(1)

pub fn redeposit( origin: OriginFor<T>, collection: T::CollectionId, items: Vec<T::ItemId> ) -> DispatchResult

Re-evaluate the deposits on some items.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection of the items to be reevaluated.

items: The items of the collection whose deposits will be reevaluated.

NOTE: This exists as a best-effort function. Any items which are unknown or in the case that the owner account does not have reservable funds to pay for a deposit increase are ignored. Generally the owner isn’t going to call this on items whose existing deposit is less than the refreshed deposit as it would only cost them, so it’s of little consequence.

It will still return an error in the case that the collection is unknown or the signer is not permitted to call it.

Weight: O(items.len())

pub fn lock_item_transfer( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId ) -> DispatchResult

Disallow further unprivileged transfer of an item.

Origin must be Signed and the sender should be the Freezer of the collection.

collection: The collection of the item to be changed.

item: The item to become non-transferable.

Emits ItemTransferLocked.

Weight: O(1)

pub fn unlock_item_transfer( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId ) -> DispatchResult

Re-allow unprivileged transfer of an item.

Origin must be Signed and the sender should be the Freezer of the collection.

collection: The collection of the item to be changed.

item: The item to become transferable.

Emits ItemTransferUnlocked.

Weight: O(1)

pub fn lock_collection( origin: OriginFor<T>, collection: T::CollectionId, lock_settings: CollectionSettings ) -> DispatchResult

Disallows specified settings for the whole collection.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection to be locked.

lock_settings: The settings to be locked.

Note: it’s possible to only lock(set) the setting, but not to unset it.

Emits CollectionLocked.

Weight: O(1)

pub fn transfer_ownership( origin: OriginFor<T>, collection: T::CollectionId, owner: <<T as SystemConfig>::Lookup as StaticLookup>::Source ) -> DispatchResult

Change the Owner of a collection.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection whose owner should be changed.

owner: The new Owner of this collection. They must have called set_accept_ownership with collection in order for this operation to succeed.

Emits OwnerChanged.

Weight: O(1)

pub fn set_team( origin: OriginFor<T>, collection: T::CollectionId, issuer: Option<<<T as SystemConfig>::Lookup as StaticLookup>::Source>, admin: Option<<<T as SystemConfig>::Lookup as StaticLookup>::Source>, freezer: Option<<<T as SystemConfig>::Lookup as StaticLookup>::Source> ) -> DispatchResult

Change the Issuer, Admin and Freezer of a collection.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the collection.

Note: by setting the role to None only the ForceOrigin will be able to change it after to Some(account).

collection: The collection whose team should be changed.

issuer: The new Issuer of this collection.

admin: The new Admin of this collection.

freezer: The new Freezer of this collection.

Emits TeamChanged.

Weight: O(1)

pub fn force_collection_owner( origin: OriginFor<T>, collection: T::CollectionId, owner: <<T as SystemConfig>::Lookup as StaticLookup>::Source ) -> DispatchResult

Change the Owner of a collection.

Origin must be ForceOrigin.

collection: The identifier of the collection.

owner: The new Owner of this collection.

Emits OwnerChanged.

Weight: O(1)

pub fn force_collection_config( origin: OriginFor<T>, collection: T::CollectionId, config: CollectionConfig<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, <T as Config<I>>::CollectionId> ) -> DispatchResult

Change the config of a collection.

Origin must be ForceOrigin.

collection: The identifier of the collection.

config: The new config of this collection.

Emits CollectionConfigChanged.

Weight: O(1)

pub fn approve_transfer( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, delegate: <<T as SystemConfig>::Lookup as StaticLookup>::Source, maybe_deadline: Option<BlockNumberFor<T>> ) -> DispatchResult

Approve an item to be transferred by a delegated third-party account.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the item.

collection: The collection of the item to be approved for delegated transfer.

item: The item to be approved for delegated transfer.

delegate: The account to delegate permission to transfer the item.

maybe_deadline: Optional deadline for the approval. Specified by providing the number of blocks after which the approval will expire

Emits TransferApproved on success.

Weight: O(1)

pub fn cancel_approval( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, delegate: <<T as SystemConfig>::Lookup as StaticLookup>::Source ) -> DispatchResult

Cancel one of the transfer approvals for a specific item.

Origin must be either:

the Force origin;

Signed with the signer being the Owner of the item;

Arguments:

collection: The collection of the item of whose approval will be cancelled.

item: The item of the collection of whose approval will be cancelled.

delegate: The account that is going to loose their approval.

Emits ApprovalCancelled on success.

Weight: O(1)

pub fn clear_all_transfer_approvals( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId ) -> DispatchResult

Cancel all the approvals of a specific item.

Origin must be either:

the Force origin;

Signed with the signer being the Owner of the item;

Arguments:

collection: The collection of the item of whose approvals will be cleared.

item: The item of the collection of whose approvals will be cleared.

Emits AllApprovalsCancelled on success.

Weight: O(1)

pub fn lock_item_properties( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, lock_metadata: bool, lock_attributes: bool ) -> DispatchResult

Disallows changing the metadata or attributes of the item.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

collection: The collection if the item.

item: An item to be locked.

lock_metadata: Specifies whether the metadata should be locked.

lock_attributes: Specifies whether the attributes in the CollectionOwner namespace should be locked.

Note: lock_attributes affects the attributes in the CollectionOwner namespace only. When the metadata or attributes are locked, it won’t be possible the unlock them.

Emits ItemPropertiesLocked.

Weight: O(1)

pub fn set_attribute( origin: OriginFor<T>, collection: T::CollectionId, maybe_item: Option<T::ItemId>, namespace: AttributeNamespace<T::AccountId>, key: BoundedVec<u8, T::KeyLimit>, value: BoundedVec<u8, T::ValueLimit> ) -> DispatchResult

Set an attribute for a collection or item.

Origin must be Signed and must conform to the namespace ruleset:

CollectionOwner namespace could be modified by the collection Admin only;

ItemOwner namespace could be modified by the maybe_item owner only. maybe_item should be set in that case;

Account(AccountId) namespace could be modified only when the origin was given a permission to do so;

The funds of origin are reserved according to the formula: AttributeDepositBase + DepositPerByte * (key.len + value.len) taking into account any already reserved funds.

collection: The identifier of the collection whose item’s metadata to set.

maybe_item: The identifier of the item whose metadata to set.

namespace: Attribute’s namespace.

key: The key of the attribute.

value: The value to which to set the attribute.

Emits AttributeSet.

Weight: O(1)

pub fn force_set_attribute( origin: OriginFor<T>, set_as: Option<T::AccountId>, collection: T::CollectionId, maybe_item: Option<T::ItemId>, namespace: AttributeNamespace<T::AccountId>, key: BoundedVec<u8, T::KeyLimit>, value: BoundedVec<u8, T::ValueLimit> ) -> DispatchResult

Force-set an attribute for a collection or item.

Origin must be ForceOrigin.

If the attribute already exists and it was set by another account, the deposit will be returned to the previous owner.

set_as: An optional owner of the attribute.

collection: The identifier of the collection whose item’s metadata to set.

maybe_item: The identifier of the item whose metadata to set.

namespace: Attribute’s namespace.

key: The key of the attribute.

value: The value to which to set the attribute.

Emits AttributeSet.

Weight: O(1)

pub fn clear_attribute( origin: OriginFor<T>, collection: T::CollectionId, maybe_item: Option<T::ItemId>, namespace: AttributeNamespace<T::AccountId>, key: BoundedVec<u8, T::KeyLimit> ) -> DispatchResult

Clear an attribute for a collection or item.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the attribute.

Any deposit is freed for the collection’s owner.

collection: The identifier of the collection whose item’s metadata to clear.

maybe_item: The identifier of the item whose metadata to clear.

namespace: Attribute’s namespace.

key: The key of the attribute.

Emits AttributeCleared.

Weight: O(1)

pub fn approve_item_attributes( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, delegate: <<T as SystemConfig>::Lookup as StaticLookup>::Source ) -> DispatchResult

Approve item’s attributes to be changed by a delegated third-party account.

Origin must be Signed and must be an owner of the item.

collection: A collection of the item.

item: The item that holds attributes.

delegate: The account to delegate permission to change attributes of the item.

Emits ItemAttributesApprovalAdded on success.

pub fn cancel_item_attributes_approval( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, delegate: <<T as SystemConfig>::Lookup as StaticLookup>::Source, witness: CancelAttributesApprovalWitness ) -> DispatchResult

Cancel the previously provided approval to change item’s attributes. All the previously set attributes by the delegate will be removed.

Origin must be Signed and must be an owner of the item.

collection: Collection that the item is contained within.

item: The item that holds attributes.

delegate: The previously approved account to remove.

Emits ItemAttributesApprovalRemoved on success.

pub fn set_metadata( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, data: BoundedVec<u8, T::StringLimit> ) -> DispatchResult

Set the metadata for an item.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

If the origin is Signed, then funds of signer are reserved according to the formula: MetadataDepositBase + DepositPerByte * data.len taking into account any already reserved funds.

collection: The identifier of the collection whose item’s metadata to set.

item: The identifier of the item whose metadata to set.

data: The general information of this item. Limited in length by StringLimit.

Emits ItemMetadataSet.

Weight: O(1)

pub fn clear_metadata( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId ) -> DispatchResult

Clear the metadata for an item.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

Any deposit is freed for the collection’s owner.

collection: The identifier of the collection whose item’s metadata to clear.

item: The identifier of the item whose metadata to clear.

Emits ItemMetadataCleared.

Weight: O(1)

pub fn set_collection_metadata( origin: OriginFor<T>, collection: T::CollectionId, data: BoundedVec<u8, T::StringLimit> ) -> DispatchResult

Set the metadata for a collection.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

If the origin is Signed, then funds of signer are reserved according to the formula: MetadataDepositBase + DepositPerByte * data.len taking into account any already reserved funds.

collection: The identifier of the item whose metadata to update.

data: The general information of this item. Limited in length by StringLimit.

Emits CollectionMetadataSet.

Weight: O(1)

pub fn clear_collection_metadata( origin: OriginFor<T>, collection: T::CollectionId ) -> DispatchResult

Clear the metadata for a collection.

Origin must be either ForceOrigin or Signed and the sender should be the Admin of the collection.

Any deposit is freed for the collection’s owner.

collection: The identifier of the collection whose metadata to clear.

Emits CollectionMetadataCleared.

Weight: O(1)

pub fn set_accept_ownership( origin: OriginFor<T>, maybe_collection: Option<T::CollectionId> ) -> DispatchResult

Set (or reset) the acceptance of ownership for a particular account.

Origin must be Signed and if maybe_collection is Some, then the signer must have a provider reference.

maybe_collection: The identifier of the collection whose ownership the signer is willing to accept, or if None, an indication that the signer is willing to accept no ownership transferal.

Emits OwnershipAcceptanceChanged.

pub fn set_collection_max_supply( origin: OriginFor<T>, collection: T::CollectionId, max_supply: u32 ) -> DispatchResult

Set the maximum number of items a collection could have.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the collection.

collection: The identifier of the collection to change.

max_supply: The maximum number of items a collection could have.

Emits CollectionMaxSupplySet event when successful.

pub fn update_mint_settings( origin: OriginFor<T>, collection: T::CollectionId, mint_settings: MintSettings<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, BlockNumberFor<T>, T::CollectionId> ) -> DispatchResult

Update mint settings.

Origin must be either ForceOrigin or Signed and the sender should be the Issuer of the collection.

collection: The identifier of the collection to change.

mint_settings: The new mint settings.

Emits CollectionMintSettingsUpdated event when successful.

pub fn set_price( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, price: Option<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>, whitelisted_buyer: Option<<<T as SystemConfig>::Lookup as StaticLookup>::Source> ) -> DispatchResult

Set (or reset) the price for an item.

Origin must be Signed and must be the owner of the item.

collection: The collection of the item.

item: The item to set the price for.

price: The price for the item. Pass None, to reset the price.

buyer: Restricts the buy operation to a specific account.

Emits ItemPriceSet on success if the price is not None. Emits ItemPriceRemoved on success if the price is None.

pub fn buy_item( origin: OriginFor<T>, collection: T::CollectionId, item: T::ItemId, bid_price: <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance ) -> DispatchResult

Allows to buy an item if it’s up for sale.

Origin must be Signed and must not be the owner of the item.

collection: The collection of the item.

item: The item the sender wants to buy.

bid_price: The price the sender is willing to pay.

Emits ItemBought on success.

pub fn pay_tips( origin: OriginFor<T>, tips: BoundedVec<ItemTip<<T as Config<I>>::CollectionId, <T as Config<I>>::ItemId, <T as SystemConfig>::AccountId, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>, T::MaxTips> ) -> DispatchResult

Allows to pay the tips.

Origin must be Signed.

tips: Tips array.

Emits TipSent on every tip transfer.

pub fn create_swap( origin: OriginFor<T>, offered_collection: T::CollectionId, offered_item: T::ItemId, desired_collection: T::CollectionId, maybe_desired_item: Option<T::ItemId>, maybe_price: Option<PriceWithDirection<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>>, duration: BlockNumberFor<T> ) -> DispatchResult

Register a new atomic swap, declaring an intention to send an item in exchange for desired_item from origin to target on the current blockchain. The target can execute the swap during the specified duration of blocks (if set). Additionally, the price could be set for the desired item.

Origin must be Signed and must be an owner of the item.

collection: The collection of the item.

item: The item an owner wants to give.

desired_collection: The collection of the desired item.

desired_item: The desired item an owner wants to receive.

maybe_price: The price an owner is willing to pay or receive for the desired item.

duration: A deadline for the swap. Specified by providing the number of blocks after which the swap will expire.

Emits SwapCreated on success.

pub fn cancel_swap( origin: OriginFor<T>, offered_collection: T::CollectionId, offered_item: T::ItemId ) -> DispatchResult

Cancel an atomic swap.

Origin must be Signed. Origin must be an owner of the item if the deadline hasn’t expired.

collection: The collection of the item.

item: The item an owner wants to give.

Emits SwapCancelled on success.

pub fn claim_swap( origin: OriginFor<T>, send_collection: T::CollectionId, send_item: T::ItemId, receive_collection: T::CollectionId, receive_item: T::ItemId, witness_price: Option<PriceWithDirection<<<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>> ) -> DispatchResult

Claim an atomic swap. This method executes a pending swap, that was created by a counterpart before.

Origin must be Signed and must be an owner of the item.

send_collection: The collection of the item to be sent.

send_item: The item to be sent.

receive_collection: The collection of the item to be received.

receive_item: The item to be received.

witness_price: A price that was previously agreed on.

Emits SwapClaimed on success.

pub fn mint_pre_signed( origin: OriginFor<T>, mint_data: Box<PreSignedMint<<T as Config<I>>::CollectionId, <T as Config<I>>::ItemId, <T as SystemConfig>::AccountId, BlockNumberFor<T>, <<T as Config<I>>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>>, signature: T::OffchainSignature, signer: T::AccountId ) -> DispatchResult

Mint an item by providing the pre-signed approval.

Origin must be Signed.

mint_data: The pre-signed approval that consists of the information about the item, its metadata, attributes, who can mint it (None for anyone) and until what block number.

signature: The signature of the data object.

signer: The data object’s signer. Should be an Issuer of the collection.

Emits Issued on success. Emits AttributeSet if the attributes were provided. Emits ItemMetadataSet if the metadata was not empty.

pub fn set_attributes_pre_signed( origin: OriginFor<T>, data: PreSignedAttributes<<T as Config<I>>::CollectionId, <T as Config<I>>::ItemId, <T as SystemConfig>::AccountId, BlockNumberFor<T>>, signature: T::OffchainSignature, signer: T::AccountId ) -> DispatchResult

Set attributes for an item by providing the pre-signed approval.

Origin must be Signed and must be an owner of the data.item.

data: The pre-signed approval that consists of the information about the item, attributes to update and until what block number.

signature: The signature of the data object.

signer: The data object’s signer. Should be an Admin of the collection for the CollectionOwner namespace.

Emits AttributeSet for each provided attribute. Emits ItemAttributesApprovalAdded if the approval wasn’t set before. Emits PreSignedAttributesSet on success.