pallet_nfts

Type Alias CollectionConfigFor

pub type CollectionConfigFor<T, I = ()> = CollectionConfig<BalanceOf<T, I>, BlockNumberFor<T, I>, <T as Config<I>>::CollectionId>;

A type alias for the settings configuration of a collection.

Aliased Type

pub struct CollectionConfigFor<T, I = ()> {
    pub settings: CollectionSettings,
    pub max_supply: Option<u32>,
    pub mint_settings: MintSettings<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <<T as Config<I>>::BlockNumberProvider as BlockNumberProvider>::BlockNumber, <T as Config<I>>::CollectionId>,
}

Fields
settings: CollectionSettings

Collection’s settings.

max_supply: Option<u32>

Collection’s max supply.

mint_settings: MintSettings<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <<T as Config<I>>::BlockNumberProvider as BlockNumberProvider>::BlockNumber, <T as Config<I>>::CollectionId>

Default settings each item will get during the mint.