pallet_nfts

Type Alias CollectionDetailsFor

pub type CollectionDetailsFor<T, I> = CollectionDetails<<T as SystemConfig>::AccountId, DepositBalanceOf<T, I>>;

A type alias representing the details of a collection.

Aliased Type

pub struct CollectionDetailsFor<T, I> {
    pub owner: <T as Config>::AccountId,
    pub owner_deposit: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    pub items: u32,
    pub item_metadatas: u32,
    pub item_configs: u32,
    pub attributes: u32,
}

Fields
owner: <T as Config>::AccountId

Collection’s owner.

owner_deposit: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The total balance deposited by the owner for all the storage data associated with this collection. Used by destroy.

items: u32

The total number of outstanding items of this collection.

item_metadatas: u32

The total number of outstanding item metadata of this collection.

item_configs: u32

The total number of outstanding item configs of this collection.

attributes: u32

The total number of attributes for this collection.