pallet_nfts

Type Alias ItemDetailsFor

pub type ItemDetailsFor<T, I> = ItemDetails<<T as SystemConfig>::AccountId, ItemDepositOf<T, I>, ApprovalsOf<T, I>>;

A type that holds the details of a single item.

Aliased Type

pub struct ItemDetailsFor<T, I> {
    pub owner: <T as Config>::AccountId,
    pub approvals: BoundedBTreeMap<<T as Config>::AccountId, Option<<<T as Config<I>>::BlockNumberProvider as BlockNumberProvider>::BlockNumber>, <T as Config<I>>::ApprovalsLimit>,
    pub deposit: ItemDeposit<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <T as Config>::AccountId>,
}

Fields
owner: <T as Config>::AccountId

The owner of this item.

approvals: BoundedBTreeMap<<T as Config>::AccountId, Option<<<T as Config<I>>::BlockNumberProvider as BlockNumberProvider>::BlockNumber>, <T as Config<I>>::ApprovalsLimit>

The approved transferrer of this item, if one is set.

deposit: ItemDeposit<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <T as Config>::AccountId>

The amount held in the pallet’s default account for this item. Free-hold items will have this as zero.