Crate pallet_nfts

Nfts Module

A simple, secure module for dealing with non-fungible items.

Related Modules

System

Support

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

macros

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_nfts

Macros

_runtime_benchmarks_enabled

runtime_benchmarks_enabled

Enable/disable the given code depending on feature = "runtime-benchmarks" being enabled for the crate or not.

Structs

AttributeDeposit

Information about the reserved attribute deposit.

CancelAttributesApprovalWitness

A witness data to cancel attributes approval operation.

CollectionConfig

Collection’s configuration.

CollectionDetails

Information about a collection.

CollectionMetadata

Information about the collection’s metadata.

CollectionRoles

A wrapper type that implements Codec.

CollectionSettings

Wrapper type for BitFlags<CollectionSetting> that implements Codec.

DestroyWitness

Witness data for the destroy transactions.

ItemConfig

Item’s configuration.

ItemDeposit

Information about the reserved item deposit.

ItemDetails

Information concerning the ownership of a single unique item.

ItemMetadata

Information about the item’s metadata.

ItemMetadataDeposit

Information about the reserved item’s metadata deposit.

ItemSettings

Wrapper type for BitFlags<ItemSetting> that implements Codec.

ItemTip

Information about the tip.

MintSettings

Holds the information about minting.

MintWitness

Witness data for items mint transactions.

PalletFeatures

Wrapper type for BitFlags<PalletFeature> that implements Codec.

PendingSwap

Information about the pending swap.

PreSignedAttributes

PreSignedMint

PriceWithDirection

Holds the details about the price.

Enums

AttributeNamespace

Attribute namespaces for non-fungible tokens.

CollectionRole

Support for up to 8 different roles for collections.

CollectionSetting

Support for up to 64 user-enabled features on a collection.

ItemSetting

Support for up to 64 user-enabled features on an item.

MintType

Mint type. Can the NFT be create by anyone, or only the creator of the collection, or only by wallets that already hold an NFT from a certain collection? The ownership of a privately minted NFT is still publicly visible.

PalletAttributes

A list of possible pallet-level attributes.

PalletFeature

Support for up to 64 system-enabled features on a collection.

PriceDirection

Specifies whether the tokens will be sent or received.

Constants

LOG_TARGET

The log target of this pallet.

Type Aliases

ApprovalsOf

A type alias for keeping track of approvals used by a single item.

AttributeDepositOf

A type that holds the deposit amount for an item’s attribute.

BalanceOf

A type alias for an accounts balance.

CollectionConfigFor

A type alias for the settings configuration of a collection.

CollectionDetailsFor

A type alias representing the details of a collection.

DepositBalanceOf

A type alias for handling balance deposits.

ItemAttributesApprovals

A type alias for keeping track of approvals for an item’s attributes.

ItemDepositOf

A type that holds the deposit for a single item.

ItemDetailsFor

A type that holds the details of a single item.

ItemMetadataDepositOf

A type that holds the deposit amount for an item’s metadata.

ItemPrice

A type alias to represent the price of an item.

ItemTipOf

A type alias for the tips held by a single item.

PreSignedAttributesOf

A type alias for the pre-signed minting configuration on the attribute level of an item.

PreSignedMintOf

A type alias for the pre-signed minting configuration for a specified collection.