Module pallet_nfts::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

approve_item_attributes

Approve item’s attributes to be changed by a delegated third-party account.

approve_transfer

Approve an item to be transferred by a delegated third-party account.

burn

Destroy a single item.

buy_item

Allows to buy an item if it’s up for sale.

cancel_approval

Cancel one of the transfer approvals for a specific item.

cancel_item_attributes_approval

Cancel the previously provided approval to change item’s attributes. All the previously set attributes by the delegate will be removed.

cancel_swap

Cancel an atomic swap.

claim_swap

Claim an atomic swap. This method executes a pending swap, that was created by a counterpart before.

clear_all_transfer_approvals

Cancel all the approvals of a specific item.

clear_attribute

Clear an attribute for a collection or item.

clear_collection_metadata

Clear the metadata for a collection.

clear_metadata

Clear the metadata for an item.

create

Issue a new collection of non-fungible items from a public origin.

create_swap

Register a new atomic swap, declaring an intention to send an item in exchange for desired_item from origin to target on the current blockchain. The target can execute the swap during the specified duration of blocks (if set). Additionally, the price could be set for the desired item.

destroy

Destroy a collection of fungible items.

force_collection_config

Change the config of a collection.

force_collection_owner

Change the Owner of a collection.

force_create

Issue a new collection of non-fungible items from a privileged origin.

force_mint

Mint an item of a particular collection from a privileged origin.

force_set_attribute

Force-set an attribute for a collection or item.

lock_collection

Disallows specified settings for the whole collection.

lock_item_properties

Disallows changing the metadata or attributes of the item.

lock_item_transfer

Disallow further unprivileged transfer of an item.

mint

Mint an item of a particular collection.

mint_pre_signed

Mint an item by providing the pre-signed approval.

pay_tips

Allows to pay the tips.

redeposit

Re-evaluate the deposits on some items.

set_accept_ownership

Set (or reset) the acceptance of ownership for a particular account.

set_attribute

Set an attribute for a collection or item.

set_attributes_pre_signed

Set attributes for an item by providing the pre-signed approval.

set_collection_max_supply

Set the maximum number of items a collection could have.

set_collection_metadata

Set the metadata for a collection.

set_metadata

Set the metadata for an item.

set_price

Set (or reset) the price for an item.

set_team

Change the Issuer, Admin and Freezer of a collection.

transfer

Move an item from the sender account to another.

transfer_ownership

Change the Owner of a collection.

unlock_item_transfer

Re-allow unprivileged transfer of an item.

update_mint_settings

Update mint settings.