Function pallet_nfts::pallet::dispatchables::force_set_attribute

pub fn force_set_attribute<T: Config<I>, I: 'static>(
    set_as: Option<T::AccountId>,
    collection: T::CollectionId,
    maybe_item: Option<T::ItemId>,
    namespace: AttributeNamespace<T::AccountId>,
    key: BoundedVec<u8, T::KeyLimit>,
    value: BoundedVec<u8, T::ValueLimit>,
)

Force-set an attribute for a collection or item.

Origin must be ForceOrigin.

If the attribute already exists and it was set by another account, the deposit will be returned to the previous owner.

set_as: An optional owner of the attribute.

collection: The identifier of the collection whose item’s metadata to set.

maybe_item: The identifier of the item whose metadata to set.

namespace: Attribute’s namespace.

key: The key of the attribute.

value: The value to which to set the attribute.

Emits AttributeSet.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::force_set_attribute.